refactor to move modal and associated logic to WodinApp
