+++ app/static/src/app/components/WodinApp.vue
@@ -1,4 +1,7 @@
 <template>
+    <session-initialise-modal v-if="!sessionInitialised && !immediateInitialise"
+                            @new-session="newSession"
+                            @reload-session="reloadSession"></session-initialise-modal>
     <div class="container">
         <div class="row">
             <div class="col-12">
@@ -21,8 +24,12 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent } from "vue";
+import {
+    computed, defineComponent
+} from "vue";
 import { useStore } from "vuex";
+import SessionInitialiseModal from "./SessionInitialiseModal.vue";
+import { AppStateAction } from "../store/appState/actions";
 import ErrorsAlert from "./ErrorsAlert.vue";
 import WodinPanels from "./WodinPanels.vue";
 import LoadingSpinner from "./LoadingSpinner.vue";
@@ -30,6 +37,7 @@ import LoadingSpinner from "./LoadingSpinner.vue";
 export default defineComponent({
     name: "WodinApp",
     components: {
+        SessionInitialiseModal,
         LoadingSpinner,
         ErrorsAlert,
         WodinPanels
@@ -40,9 +48,38 @@ export default defineComponent({
         const appType = computed(() => store.state.appType);
         const loading = computed(() => !store.state.config);
 
+        const sessionInitialised = computed(() => store.state.configured);
+        const latestSessionId = computed(() => store.state.sessions.latestSessionId);
+        const loadSessionId = computed(() => store.state.loadSessionId);
+
+        const initialise = (sessionId: string, copySession = true) => {
+            store.dispatch(AppStateAction.InitialiseSession, { loadSessionId: sessionId, copySession });
+        };
+
+        // If there's a load (share) session id, then immediately initialise with that
+        // If there is no latest session id then immediately initialise a new session - display modal conditional on
+        // this value as well as sessionInitialised to avoid momentary display of modal when this is true.
+        const immediateInitialise = computed(() => loadSessionId.value || !latestSessionId.value);
+
+        if (immediateInitialise.value) {
+            initialise(loadSessionId.value || "");
+        }
+
+        const newSession = () => {
+            initialise("");
+        };
+
+        const reloadSession = () => {
+            initialise(latestSessionId.value!, false);
+        };
+
         return {
             appType,
-            loading
+            loading,
+            sessionInitialised,
+            immediateInitialise,
+            newSession,
+            reloadSession
         };
     }
 });
